quantity

+++ app/Http/Controllers/CartController.php
@@ -16,34 +16,55 @@ class CartController extends Controller
             $carts = auth()->user()->cart->pluck('product_id');
             $product = Product::wherein('id',$carts)->get();
             
+            
             return view('cart',compact('cart','product'));
         }
         else{
             return redirect()->route('home');
         }
     }
-    
+    // public function available($request){
+    //  return 
+    // }
     public function store(Request $request){
         $user_id=auth()->user()->id;
         $product_id=$request->product_id;
+        
         Cart::create([
             'user_id' => $user_id,
             'product_id' => $product_id,
-        ]);
-        return redirect()->route('home');
+            
+            ]);
+            return redirect()->route('home');
+         
     }
     public function buy(Request $request){
         
         $user =auth()->user()->id;
+        $product_id=$request->product;
         
+        $quantity = Product::where('id',$product_id)->pluck('quantity');
+        
+        foreach($quantity as $quantity){   ///cause quantity is a fucking array and need as a int
+
+       if($quantity>=$request->quantity){
         Buy::create([
             'user_id' => $user,
             'product_id' => $request->product,
+            'quantity' =>$request->quantity,
+            
         ]);
         $boughtproduct = Cart::where('product_id',$request->product)->where('user_id',$user);
         $boughtproduct->delete();
-        //Cart::update();     to do after quantity this is just firstphase
         return redirect('/cart');
 
     }
+    else{
+        $message= "please order less";
+        $product = Product::where('id',$product_id)->get();
+        
+        return redirect('/cart');
+    }
+}
+}
 }
